+++ crates/delta/src/routes/webhooks/mod.rs
@@ -2,16 +2,22 @@ use rocket::Route;
 use revolt_rocket_okapi::revolt_okapi::openapi3::OpenApi;
 
 mod webhook_delete;
+mod webhook_delete_token;
 mod webhook_edit;
+mod webhook_edit_token;
 mod webhook_execute;
+mod webhook_fetch_token;
 mod webhook_fetch;
 mod webhook_execute_github;
 
 pub fn routes() -> (Vec<Route>, OpenApi) {
     openapi_get_routes_spec![
+        webhook_delete_token::req,
         webhook_delete::req,
+        webhook_edit_token::req,
         webhook_edit::req,
         webhook_execute::req,
+        webhook_fetch_token::req,
         webhook_fetch::req,
         webhook_execute_github::req
     ]
